# contracts/v2/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "BarRevisionPolicy": {
        "enum": [
          "UNSPECIFIED",
          "LATEST",
          "INITIAL_ONLY",
          "EMIT_REVISIONS"
        ],
        "title": "BarRevisionPolicy",
        "type": "string"
      },
...
      "BatchRequirementModel": {
        "additionalProperties": false,
        "properties": {
          "consumer_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Consumer Id",
            "type": "string"
          },
          "require_all": {
            "default": true,
            "title": "Require All",
            "type": "boolean"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementModel"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Requirements",
            "type": "array"
          }
        },
        "required": [
          "consumer_id",
          "requirements"
        ],
        "title": "BatchRequirementModel",
        "type": "object"
      },
      "BatchResponse": {
        "additionalProperties": false,
        "properties": {
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "partial": {
            "title": "Partial",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BatchItemResponse"
            },
            "title": "Results",
            "type": "array"
          },
          "schema": {
            "default": "qdl.marketdata.batch.v2",
            "title": "Schema",
            "type": "string"
          },
          "success_count": {
            "title": "Success Count",
            "type": "integer"
          }
        },
        "required": [
          "request_id",
          "partial",
          "success_count",
          "error_count",
          "results"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
...
      "ConsumerGrade": {
        "enum": [
          "UNSPECIFIED",
          "EXECUTION",
          "ALPHA",
          "RESEARCH"
        ],
        "title": "ConsumerGrade",
        "type": "string"
      },
...
      "FeedType": {
        "enum": [
          "UNSPECIFIED",
          "TRADE",
          "QUOTE",
          "BAR",
          "BOOK_SNAPSHOT",
          "BOOK_DELTA",
          "FUNDING_RATE",
          "OPEN_INTEREST",
          "MARK_INDEX_PRICE",
          "TICKER"
        ],
        "title": "FeedType",
        "type": "string"
      },
...
      "GapPolicy": {
        "enum": [
          "UNSPECIFIED",
          "BLOCK",
          "PAUSE",
          "OBSERVE"
        ],
        "title": "GapPolicy",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InstrumentPageResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstrumentView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "schema": {
            "default": "qdl.instruments.page.v2",
            "title": "Schema",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "title": "InstrumentPageResponse",
        "type": "object"
      },
      "InstrumentResponse": {
        "additionalProperties": false,
        "properties": {
          "asset_class": {
            "title": "Asset Class",
            "type": "string"
          },
          "canonical_symbol": {
            "title": "Canonical Symbol",
            "type": "string"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "instrument_uid": {
            "title": "Instrument Uid",
            "type": "string"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "metadata_revision": {
            "minimum": 1.0,
            "title": "Metadata Revision",
            "type": "integer"
          },
          "native_symbol": {
            "title": "Native Symbol",
            "type": "string"
          },
          "product_type": {
            "title": "Product Type",
            "type": "string"
          },
          "schema": {
            "default": "qdl.instrument.v2",
            "title": "Schema",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "venue": {
            "title": "Venue",
            "type": "string"
          }
        },
        "required": [
          "instrument_uid",
          "instrument_id",
          "venue",
          "market",
          "product_type",
          "canonical_symbol",
          "metadata_revision",
          "asset_class",
          "native_symbol",
          "status"
        ],
        "title": "InstrumentResponse",
        "type": "object"
      },
...
      "IntervalSourcePolicy": {
        "enum": [
          "NATIVE_ONLY",
          "NATIVE_OR_EXACT_RESAMPLE"
        ],
        "title": "IntervalSourcePolicy",
        "type": "string"
      },
...
      "RecoveryPolicy": {
        "enum": [
          "UNSPECIFIED",
          "SNAPSHOT_AND_REPLAY",
          "FRESH_SNAPSHOT",
          "NONE"
        ],
        "title": "RecoveryPolicy",
        "type": "string"
      },
...
      "SnapshotResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MarketDataView"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "schema": {
            "default": "qdl.marketdata.snapshot.v2",
            "title": "Schema",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "data"
        ],
        "title": "SnapshotResponse",
        "type": "object"
      },
...
      "StalePolicy": {
        "enum": [
          "UNSPECIFIED",
          "BLOCK",
          "PAUSE",
          "OBSERVE"
        ],
        "title": "StalePolicy",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "QDLConsumerIdentity": {
        "in": "header",
        "name": "X-QDL-Consumer-ID",
        "type": "apiKey"
      },
      "QDLWorkloadBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Quant Data Layer V2",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/instruments": {
      "get": {
        "operationId": "list_instruments_v2_instruments_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "QDLConsumerIdentity": [],
            "QDLWorkloadBearer": []
          }
        ],
        "summary": "List Instruments",
        "tags": [
          "market-data-v2"
        ]
      }
    },
    "/v2/instruments/{identity}": {
      "get": {
        "operationId": "get_instrument_v2_instruments__identity__get",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "QDLConsumerIdentity": [],
            "QDLWorkloadBearer": []
          }
        ],
        "summary": "Get Instrument",
        "tags": [
          "market-data-v2"
        ]
      }
    },
    "/v2/market-data/warmup:batch": {
      "post": {
        "operationId": "warmup_batch_v2_market_data_warmup_batch_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-QDL-Purpose",
            "required": true,
            "schema": {
              "title": "X-Qdl-Purpose",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequirementModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "QDLConsumerIdentity": [],
            "QDLWorkloadBearer": []
          }
        ],
        "summary": "Warmup Batch",
        "tags": [
          "market-data-v2"
        ]
      }
    },
...
    "/v2/market-data/{instrument_uid}/snapshot": {
      "get": {
        "operationId": "snapshot_v2_market_data__instrument_uid__snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_uid",
            "required": true,
            "schema": {
              "title": "Instrument Uid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "feed",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FeedType"
            }
          },
          {
            "in": "query",
            "name": "source_policy_id",
            "required": true,
            "schema": {
              "title": "Source Policy Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "consumer_grade",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ConsumerGrade",
              "default": "ALPHA"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interval"
            }
          },
          {
            "in": "query",
            "name": "max_freshness_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 86400000,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Freshness Ms"
            }
          },
          {
            "in": "query",
            "name": "require_full_coverage",
            "required": false,
            "schema": {
              "default": true,
              "title": "Require Full Coverage",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "require_final_bars",
            "required": false,
            "schema": {
              "default": true,
              "title": "Require Final Bars",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "stale_policy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/StalePolicy",
              "default": "BLOCK"
            }
          },
          {
            "in": "query",
            "name": "gap_policy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GapPolicy",
              "default": "BLOCK"
            }
          },
          {
            "in": "query",
            "name": "recovery",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RecoveryPolicy",
              "default": "SNAPSHOT_AND_REPLAY"
            }
          },
          {
            "in": "query",
            "name": "bar_revision_policy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BarRevisionPolicy",
              "default": "LATEST"
            }
          },
          {
            "in": "header",
            "name": "X-QDL-Purpose",
            "required": true,
            "schema": {
              "title": "X-Qdl-Purpose",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "QDLConsumerIdentity": [],
            "QDLWorkloadBearer": []
          }
        ],
        "summary": "Snapshot",
        "tags": [
          "market-data-v2"
        ]
      }
    },
    "/v2/market-data/{instrument_uid}/warmup": {
      "get": {
        "operationId": "warmup_v2_market_data__instrument_uid__warmup_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_uid",
            "required": true,
            "schema": {
              "title": "Instrument Uid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "feed",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FeedType"
            }
          },
          {
            "in": "query",
            "name": "source_policy_id",
            "required": true,
            "schema": {
              "title": "Source Policy Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "consumer_grade",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ConsumerGrade",
              "default": "ALPHA"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interval"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "start_time_ns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time Ns"
            }
          },
          {
            "in": "query",
            "name": "end_time_ns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time Ns"
            }
          },
          {
            "in": "query",
            "name": "interval_source_policy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IntervalSourcePolicy",
              "default": "NATIVE_OR_EXACT_RESAMPLE"
            }
          },
          {
            "in": "query",
            "name": "max_cache_age_ms",
            "required": false,
            "schema": {
              "default": 60000,
              "maximum": 86400000,
              "minimum": 0,
              "title": "Max Cache Age Ms",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "deadline_ms",
            "required": false,
            "schema": {
              "default": 20000,
              "maximum": 120000,
              "minimum": 100,
              "title": "Deadline Ms",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_freshness_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 86400000,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Freshness Ms"
            }
          },
          {
            "in": "query",
            "name": "require_full_coverage",
            "required": false,
            "schema": {
              "default": true,
              "title": "Require Full Coverage",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "require_final_bars",
            "required": false,
            "schema": {
              "default": true,
              "title": "Require Final Bars",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "stale_policy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/StalePolicy",
              "default": "BLOCK"
            }
          },
          {
            "in": "query",
            "name": "gap_policy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GapPolicy",
              "default": "BLOCK"
            }
          },
          {
            "in": "query",
            "name": "recovery",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RecoveryPolicy",
              "default": "SNAPSHOT_AND_REPLAY"
            }
          },
          {
            "in": "query",
            "name": "bar_revision_policy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BarRevisionPolicy",
              "default": "LATEST"
            }
          },
          {
            "in": "header",
            "name": "X-QDL-Purpose",
            "required": true,
            "schema": {
              "title": "X-Qdl-Purpose",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarmupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "QDLConsumerIdentity": [],
            "QDLWorkloadBearer": []
          }
        ],
        "summary": "Warmup",
        "tags": [
          "market-data-v2"
        ]
      }
    },
...
  }
}